+++ R/pas_enhanceRawData.R
@@ -52,10 +52,10 @@
 #'
 
 pas_enhanceRawData <- function(
-  pas_raw = NULL,
-  countryCodes = NULL,
-  stateCodes = NULL,
-  counties = NULL
+    pas_raw = NULL,
+    countryCodes = NULL,
+    stateCodes = NULL,
+    counties = NULL
 ) {
 
   # ----- Validate Parameters --------------------------------------------------
@@ -151,20 +151,23 @@ pas_enhanceRawData <- function(
     )) %>%
 
     # * Add core metadata -----
-    MazamaLocationUtils::table_addCoreMetadata() %>%
-
-    # TODO:  This step can be removed when MazamaLocationUtils gets upgraded to
-    # TODO:  use geohashTools to create a locationID.
-    # * Replace MLU version 0.3.8 locationID with geohash
-    dplyr::mutate(
-      locationID = MazamaCoreUtils::createLocationID(.data$longitude, .data$latitude, algorithm = "geohash")
-    ) %>%
+    # NOTE:  precision = 9 results in a precision of ~2 meters
+    MazamaLocationUtils::table_addCoreMetadata(precision = 9) %>%
 
     # Fill in new columns where possible
     dplyr::mutate(
       deviceDeploymentID = paste0(.data$locationID, "_", .data$deviceID),
       locationName = .data$name
     )
+
+  # TODO:  This step can be removed when AirMonitor gets upgraded to replace
+  # TODO:  'zip' with 'postalCode'.
+  if ( !"zip" %in% names(pas) ) {
+    pas$zip <- pas$postalCode
+  }
+  if ( !"postalCode" %in% names(pas) ) {
+    pas$postalCode <- pas$zip
+  }
 
   # Put 'deviceDeploymentID' and 'deviceID' in front
   startingIDs <- c("deviceDeploymentID", "deviceID", "locationID")